+++ openquake/commonlib/source_reader.py
@@ -61,31 +61,6 @@ def random_filtered_sources(sources, srcfilter, seed):
     return []
 
 
-def check_nonparametric_sources(fname, smodel, investigation_time):
-    """
-    :param fname:
-        full path to a source model file
-    :param smodel:
-        source model object
-    :param investigation_time:
-        investigation_time to compare with in the case of
-        nonparametric sources
-    :returns:
-        the nonparametric sources in the model
-    :raises:
-        a ValueError if the investigation_time is different from the expected
-    """
-    # NonParametricSeismicSources
-    np = [src for sg in smodel.src_groups for src in sg
-          if hasattr(src, 'data')]
-    if np and smodel.investigation_time != investigation_time:
-        raise ValueError(
-            'The source model %s contains an investigation_time '
-            'of %s, while the job.ini has %s' % (
-                fname, smodel.investigation_time, investigation_time))
-    return np
-
-
 class SourceReader(object):
     """
     :param converter: a SourceConverter instance
@@ -105,9 +80,7 @@ class SourceReader(object):
         mags = set()
         src_groups = []
         [sm] = nrml.read_source_models([fname], self.converter, monitor)
-        check_nonparametric_sources(
-            fname, sm, self.converter.investigation_time)
-        newsm = apply_unc(path, sm)
+        newsm = apply_unc(path, sm, self.converter, monitor)
         fname_hits[fname] += 1
         for sg in newsm:
             # sample a source for each group
