+++ openquake/hazardlib/nrml.py
@@ -80,7 +80,7 @@ import collections.abc
 
 import numpy
 
-from openquake.baselib import hdf5
+from openquake.baselib import hdf5, performance
 from openquake.baselib.general import CallableDict, groupby
 from openquake.baselib.node import (
     node_to_xml, Node, striptag, ValidatingXmlParser, floatformat)
@@ -306,7 +306,32 @@ validators = {
 }
 
 
-def read_source_models(fnames, converter, monitor):
+def check_nonparametric_sources(fname, smodel, investigation_time):
+    """
+    :param fname:
+        full path to a source model file
+    :param smodel:
+        source model object
+    :param investigation_time:
+        investigation_time to compare with in the case of
+        nonparametric sources
+    :returns:
+        the nonparametric sources in the model
+    :raises:
+        a ValueError if the investigation_time is different from the expected
+    """
+    # NonParametricSeismicSources
+    np = [src for sg in smodel.src_groups for src in sg
+          if hasattr(src, 'data')]
+    if np and smodel.investigation_time != investigation_time:
+        raise ValueError(
+            'The source model %s contains an investigation_time '
+            'of %s, while the job.ini has %s' % (
+                fname, smodel.investigation_time, investigation_time))
+    return np
+
+
+def read_source_models(fnames, converter, monitor=performance.Monitor()):
     """
     :param fnames:
         list of source model files
@@ -323,6 +348,7 @@ def read_source_models(fnames, converter, monitor):
         else:
             raise ValueError('Unrecognized extension in %s' % fname)
         sm.fname = fname
+        check_nonparametric_sources(fname, sm, converter.investigation_time)
         yield sm
