+++ openquake/qa_tests_data/event_based/case_25/common2.xml
