+++ openquake/qa_tests_data/event_based/case_25/expected/hazard_curve-PGA.csv
@@ -1,3 +1,3 @@
-#,,,,,,,,,,,,"generated_by='OpenQuake engine 3.9.0-gitfdc9c26f47', start_date='2020-02-27T16:14:17', checksum=3406892581, kind='mean', investigation_time=1.0, imt='PGA'"
+#,,,,,,,,,,,,"generated_by='OpenQuake engine 3.9.0-git40908bf21c', start_date='2020-02-28T05:33:14', checksum=1192658444, kind='mean', investigation_time=1.0, imt='PGA'"
 lon,lat,depth,poe-0.0100000,poe-0.0166810,poe-0.0278256,poe-0.0464159,poe-0.0774264,poe-0.1291550,poe-0.2154435,poe-0.3593814,poe-0.5994843,poe-1.0000000
-0.00000,0.00000,-0.10000,6.939133E-01,6.269174E-01,6.052687E-01,5.595683E-01,4.620556E-01,3.021841E-01,1.562930E-01,7.688365E-02,0.000000E+00,0.000000E+00
+0.00000,0.00000,-0.10000,7.020553E-01,6.312668E-01,5.467486E-01,4.536633E-01,3.291213E-01,2.053823E-01,7.730282E-02,1.378251E-02,0.000000E+00,0.000000E+00
