+++ openquake/qa_tests_data/event_based/case_25/job2.ini
@@ -1,0 +1,42 @@
+[general]
+
+description = common + extra
+calculation_mode = classical
+compare_with_classical = true
+individual_curves = true
+
+[geometry]
+
+sites = 0.0 0.0 -0.1
+
+[logic_tree]
+
+number_of_logic_tree_samples = 0
+
+[erf]
+
+
+# Not used in this test case:
+rupture_mesh_spacing = 1.0
+# Not used in this test case:
+width_of_mfd_bin = 1.0
+
+[site_params]
+
+reference_vs30_type = measured
+reference_vs30_value = 800.0
+reference_depth_to_2pt5km_per_sec = 2.5
+reference_depth_to_1pt0km_per_sec = 50.0
+
+[calculation]
+
+source_model_logic_tree_file = source_model_logic_tree2.xml
+gsim_logic_tree_file = gsim_logic_tree.xml
+# years
+investigation_time = 1
+ses_per_logic_tree_path = 50
+intensity_measure_types_and_levels = {"PGA": logscale(.01, 1, 10)}
+truncation_level = 2.0
+# km
+maximum_distance = 200.0
+hazard_curves_from_gmfs = true
